MB-60797: Precisely expose resident ratio to prometheus
Previously, the resident ratio would always be rounded down to an
integer. Now, the exact resident ratio is exposed to prometheus
(Grafana formats to 3 decimal places).

Reviewed-on: https://review.couchbase.org/c/kv_engine/+/206933
Tested-by: Build Bot <build@couchbase.com>

+++ engines/ep/src/vb_count_visitor.h
@@ -61,10 +61,11 @@ public:
         return numVbucket;
     }
 
-    size_t getMemResidentPer() const {
+    double getMemResidentPer() const {
         if (numItems && numItems >= nonResident) {
             size_t numResident = numItems - nonResident;
-            return size_t(numResident * 100.0) / (numItems);
+            return static_cast<double>(numResident) * 100.0 /
+                   static_cast<double>(numItems);
         }
         // Note: access-scanner depends on 100% being returned for this case
         return 100;
